ADK_VERSION_requirements.txt
# Core Dependencies - ADK Version
google-adk>=1.18.0
google-genai>=1.0.0
litellm>=1.65.5

# LLM Backend
ollama>=0.4.6

# Frontend
streamlit>=1.32.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Text processing
html2text>=2020.1.16

# Environment variables
python-dotenv>=1.0.0
